Add parent-child fuctionality for more complex objects
