+++ classes/projectile.py
@@ -1,0 +1,38 @@
+from utils import Object
+from globals import WIDTH, HEIGHT
+
+class Projectile(Object):
+
+    def __init__(self, image = 'projectile', pos = (0,0), speed = 8, health = 1, direction = -1, spin = 0, angle = 0, bounds = (WIDTH, HEIGHT), damage = 1, source = None):
+        super().__init__(image, pos, speed=speed, health=health, direction=direction, timespan=10, spin=spin, angle=angle, bounds=bounds, source=source)
+        self.damage: int = damage
+
+    def update(self):
+        self.y += self.speed*self.direction
+        if self.y <= -50 or self.y >= self.bounds[1] + 50:
+            self.kill()
+
+    def collide(self, object):
+        super().collide(object)
+        # if self.source != object and self.source != object.parent and not isinstance(object, Projectile):
+        if self.source != object and self.source != object.parent and self.source != object.source:
+        # if not self.parent_hit(object) and self.source != object.source:
+            self.alive = False
+
+    # def source_hit(self, obj):
+    #     if self.source == obj:
+    #         return True
+    #     else:
+    #         return False
+
+    # def parent_hit(self, obj):
+    #     if self.source.parent:
+    #         if self.source.parent == obj:
+    #             return True
+    #         else:
+    #             return self.source.
+        
+    # def child_hit(self, obj):
+    
+    def copy(self):
+        return Projectile( image=self.image, pos=self.pos, speed=self.speed, health=self.health, direction=self.direction, spin=self.spin, angle=self.angle, )
