+++ globals.py
@@ -1,0 +1,4 @@
+WIDTH = 1000
+HEIGHT = 800
+ASTEROIDS_SPEED = 1
+ASTEROIDS_PER_SECOND = 0.4
