+++ main.py
@@ -1,12 +1,13 @@
 import pgzrun
 import random
 from laboratory import character
-from classes import Asteroid, Background
+from classes.asteroid import Asteroid
+from classes.reflector import Reflector
 from gui import Bar
 from pygame import Color
 import sys
-from utils import asteroid_images
-from globals import WIDTH, HEIGHT
+from utils import asteroid_images, Background
+from globals import *
 
 background = Background('background2')
 
@@ -14,19 +15,14 @@ objects = [character]
 healthbar = Bar((5,HEIGHT - 20), (180,10), Color(128, 0, 0), Color(50, 50, 50), 10)
 abilitybar = Bar((5,HEIGHT - 35), (180,10), Color(0, 200, 0), Color(50, 50, 50), 10)
 cooldownbar = Bar((5,HEIGHT - 35), (180,10), Color(0, 150, 0), Color(50, 50, 50), 10, reversed = True)
-    
+
 def update_enviroment():
 
-    if random.randint(0, 150)  == 1:
+    if random.randint(0, 100) < (ASTEROIDS_PER_SECOND/60)*100:
         asteroid = Asteroid(image = random.choice(asteroid_images), 
-                            pos = (random.randint(-80,WIDTH), -30), 
-                            speed = 1,
-                            health = 10, 
-                            direction = 1, 
-                            timespan = 30, 
-                            # spin = random.randint(-20,20)/100, 
-                            spin = 0,
-                            angle = random.randint(1,360)
+                            pos = (random.randint(-80,WIDTH), -30),
+                            angle = random.randint(1,360),
+                            speed=ASTEROIDS_SPEED
                            )
         
         objects.append(asteroid)
@@ -37,10 +33,7 @@ def update_objects():
 
     for obj in objects:
         
-        if obj.collidable:
-            collided_objects = [o for o in objects if o != obj and o.collidable and obj.colliderect(o)] #Exclude self and objects with no collision
-            for collided_object in collided_objects:
-                obj.collide( collided_object )
+        obj.handle_collitions(objects)
 
         created_objects = obj.update()
         if created_objects:
