--- spaceship.py
@@ -1,8 +1,0 @@
-from pgzero.actor import Actor
-
-class CustomSpaceship(Actor):
-
-    def __init__(self, image, pos=(500, 750), **kwargs):
-        super().__init__(image, pos)
-        pass
-
